Add VisiblePositionTemplate::IsValidFor()
This method allows us to check if a position is attached to a specified
Document. It is analogous to SelectionTemplate::IsValidFor() and
SelectionForUndoStep::IsValidFor(). I'm also adding
PositionTemplate::IsValidFor() and
PositionWithAffinityTemplate::IsValidFor() to implement
VisiblePositionTemplate::IsValidFor() on top of.

For Position, we can just check IsOrphan() to determine if the position
is connected to the specified document, but for PositionInFlatTree, we
need to walk all the way to to the root, since we may be connected to
the root of a shadow tree but not actually connected to the document.

Bug: 759971
Change-Id: I04fa7b065bc894f6c29928e58083efa3f2d5315c
Reviewed-on: https://chromium-review.googlesource.com/644575
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#500541}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f496799edcce9c160708df0af17718dbed0f1ec7