Media Controls: hide duration field when changed to infinity.
The duration field will not show up when the resource is known to have
an infinite duration but if it is updated after load, the UI will not
properly update. This is refactoring both code path to be the same and
therefore always hide the duration field when the duration is known to
be not finite.

Bug: 756785
Change-Id: Iab38f61fc337f39e5b6bb712bd5f1cecc9a95f05
Reviewed-on: https://chromium-review.googlesource.com/620775
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: apacible <apacible@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#500664}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 710899b571a24886905df6b0825729cadd226633