Add images locally to image decode perf benchmark
The performance benchmark for image decoding would work if you opened
the page locally. However, it did not work as part of the test harness.
This is because it attempted to access the images which were outside of
the benchmark's root folder.

Move the images into the resources/ subdirectory so they are accessible
from within the test.

BUG=717533

Change-Id: I31dbe4e70354ea336379f062255926dab3c0a5bb
Reviewed-on: https://chromium-review.googlesource.com/657925
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Adenilson Cavalcanti <cavalcantii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#500725}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f1ed0d53d2eae0a7d03e3f413bfc311c30f7b40b