Fix crash in TextSuggestionController when DOM is changed by selectstart handler
Clusterfuzz found a crash in TextSuggestionController's helper method
ComputeRangeSurroundingCaret(). The bug is that it takes the caret position,
calls PositionInFlatTree::ComputeContainerNode() (which can return a null
pointer in certain cases), and then dereferences the null pointer (to check if
it's a text node).

The root cause is that the JavaScript selectstart handler that
SelectionController calls can cause the selection position to no longer be
valid (in particular, it can become disconnected from the rest of the document).

The fix is to add some checks in SelectionController to verify that the caret
position is still valid.

Bug: 759971
Change-Id: I7a6c0ecdcbf6aba7f4f30579070eff142fdd9a2b
Reviewed-on: https://chromium-review.googlesource.com/642115
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501027}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ded4db09901e3da7a8827336528e02170b552391