MemoryInfra: Do not call OnHeapProfilingEnabled() for all providers
To call OnHeapProfilingEnabled() on all providers, it is required to
post task for each of them. This sometimes creates issues because the
task runner may not be available when enabling profiling. So, only call
on providers that support profiling.

BUG=762994

Change-Id: Ic428f7c772e982167d3eb2260197e11a1f8685cf
Reviewed-on: https://chromium-review.googlesource.com/657894
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Siddhartha S <ssid@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501501}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 261f33bbf6cedb11de15eaaf2735bf4ea708e37c