Prevent reentrancy crash in ScrollAnchorAdjustment
CHECKs proved this to be the cause of the crash in the bug. While
there are some suggested fixes in the bug that are preferable, this
patch patch simply avoids reentering PerformScrollAnchoringAdjustments
which will prevent the crash.

Bug: 745686
Change-Id: Ic2a156cb344b6b32e5aa4d5c3164421a033c757e
Reviewed-on: https://chromium-review.googlesource.com/657600
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501549}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb6e9c67ce2694bee082ed1f82e4ea3b666ceb79