Make TextSuggestionController ignore taps while text suggestion menu is opening
There's potentially a very short window of time (or maybe longer, if the UI
thread is hung) where TextSuggestionController has told the Android browser code
to open the text suggestions menu, but it hasn't actually opened yet. If a
further tap on a suggestion/spelling marker is received in this window, the 300
ms double-tap timer will be started again, and if no further input is received
in that window to invalidate the timer, TextSuggestionController will try to
show the menu a second time, which does bad things (in particular, it may try to
create a second ActiveSuggestion marker on top of the first one, which isn't
supposed to happen).

This CL adds a check to make sure we avoid this corner case.

Bug: 672259
Change-Id: I028b28b974a8f5381c82aa1485df180192b0edec
Reviewed-on: https://chromium-review.googlesource.com/665309
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501716}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 766e933e2ce53bbc838e5da580a43b5717fa8863