Make HTMLFrameOwnerElement::PluginDisposeSuspendScope more lightweight.
This allows nearly all of the logic to be inline and only a few
instructions long. The set itself is also made a persistent collection,
as otherwise DEFINE_STATIC_LOCAL will wrap it in a Persistent<> handle,
which is an unnecessary indirection here.

Virtually all of the time, there is no suspended plugin disposal in a
child node removal.

Saves about 3% in a microbenchmark that does appendChild/removeChild
in a loop.

Bug: 764618
Change-Id: I231b8c769d801c461482a129345517303baf657c
Reviewed-on: https://chromium-review.googlesource.com/663299
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501850}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2700afca42099e3af0c2eed7a1fef7f1559c55d3