Take a LocalDOMWindow instead of a LocalFrame in Performance::Create
it's logically associated with a window rather than a frame, and this
enables graceful handling of a detached window. Performance uses
TaskRunnerHelper to look up a task runner in its constructor, and
TaskRunnerHelper is not resilient to doing that lookup for a null
LocalFrame.

BUG=763685
Test=fast/performance/detached-event-timestamp.html

Change-Id: I8e850fe55e7554c3992ba1ec449c6cf7e2b1057b
Reviewed-on: https://chromium-review.googlesource.com/666001
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Shubhie Panicker <panicker@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502035}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5232191c7f28c18cfb16bd05026aad7204ddf42a