Check selection validation after webkitBeforeTextInserted fired
InsertLineBreak and InsertParagraphSeperator dispatch
webkitBeforeTextInserted event.
If the focus nodes get removed, the nodes are disconnected.
So if the canonicalization is done just after this event
DCHECK(position.isConnected) is hit.
To avoid check if the selection is valid for the current document.

Bug: 
Change-Id: I2e278a1868fd2f6f3fe197dce70c53afaabc0c69
Reviewed-on: https://chromium-review.googlesource.com/667458
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 05e1938f03352beeb0dc27d068bae5a768fe2862