WPT export: Add code review links back to provisional PRs
crrev.com/c/658284 switched to using git fetch for in-flight CLs, which
unintentionally changed what was included in PR description. The commit
message (of an in-flight CL) does not contain the Reviewed-on footer
(the link to code review), which is useful and added back in this CL.

Besides, in this CL I re-examined the ending newlines in commit messages
and PR description to make sure the mock is consistent with real output
and there is no unintended empty new line.

Bug: 743153
Change-Id: If0f820d70bb47c0c27427d56bdbef590c71d759a
Reviewed-on: https://chromium-review.googlesource.com/669008
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Robert Ma <robertma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502420}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60c683c1d66d9ba50220845150cc27caeedf3b7a