PaymentHandler: Fix [Exposed] extended attribute in PaymentInstruments.
This CL is fixing the [Exposed=ServiceWorker] extended attribute in
PaymentInstruments. It was added in https://crrev.com/2775343002 but
it just seems to be inserted accidentally during copy-n-paste.
So, we should correct the extended attribute as per the spec[1].

This is also fixing a web-platform-test (interfaces.https.any.html).

[1] https://w3c.github.io/payment-handler/#paymentinstruments-interface

Bug: none
Change-Id: I502be6b6c1f61e5eb7277c8b52e246d2e1afc4c8
Reviewed-on: https://chromium-review.googlesource.com/669462
Commit-Queue: Jinho Bang <jinho.bang@samsung.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502631}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 388fd6b9683f6b8921308e0debf471ba2cd219d7