Clips accessible bounding box to nearest window.
For example, if a bounding box extends partially below of a window,
it will be shrunk in Y to match the border of the window. If a
bounding box is fully below a window, it will be decreased to a
height of 1 and placed at the very bottom of that window.

R=dmazzoni@chromium.org

Bug: 764065
Change-Id: Ibb105874710ebb2a0f3a7623d81b10a3f51f09ca
Reviewed-on: https://chromium-review.googlesource.com/669949
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Katie Dektar <katydek@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#502954}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77cdaac7966c31125e514b7030bf699379c844e9