Use exact path bounds for SVG shape bounding boxes
The bounding box is defined as "the tightest fitting rectangle aligned
with the axes of that element's user coordinate system that entirely
encloses it and its descendants" [1], and it directly affects the
sizing of resources using 'objectBoundingBox' units.

Hence, it is incorrect to rely on conservative/approximate path bounds
(which are not exact and usually include control points).

[1] https://www.w3.org/TR/SVG2/coords.html#BoundingBoxes

BUG=766371,230599

Change-Id: Ie3d6cce00758fa60f2af174e6cf17297998be853
Reviewed-on: https://chromium-review.googlesource.com/675184
Commit-Queue: Florin Malita <fmalita@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503478}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b7c3b90b5587cf2e76cc57fa065d5a73fcff8653