Add PlainTextRangeForEphemeralRange() helper in InputMethodControlller.cpp
There are several places in InputMethodController where we use an EphemeralRange
to create a PlainTextRange relative to the root editable element. This CL adds
a helper method that takes an EphemeralRange and returns an std::pair of the
root editable element and the resulting PlainTextRange, to allow us to reduce
code duplication.

Change-Id: I67a6c6cdad71c66426b73d3b151ab15cc0eece3e
Reviewed-on: https://chromium-review.googlesource.com/677496
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503636}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb366cedf4fecc0f0a32be170984d599dcd87e3b