Add field on SuggestionMarker to record if it's for a misspelling
Android SuggestionSpans with the FLAG_MISSPELLING flag set need to be removed
after we apply a suggestion. Therefore, to support this flag, we need to have a
type field on DocumentMarker to record whether or not the marker represents a
misspelling.

TBR=skobes@chromium.org

Bug: 672259
Change-Id: I8e2d6960b8954d3d0f680fb63345995de0930861
Reviewed-on: https://chromium-review.googlesource.com/661148
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503833}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ffbae096b1229754a33cee80abfa08fe445578e2