blink: Remove uniqueID caching for SkImages created by BitmapImage.
Currently blink caches SkImage uniqueIDs created by BitmapImage to
ensure reuse of cached decoded data for the frames of this image. The 2
systems that can cache the decode are cc and skia.

1) cc: The image decode cache in cc uses PaintImage::FrameKey to cache
the decodes and does not rely on SkImage uniqueIDs anymore.

2) skia: Skia's decode caching is tied to the lifetime of the SkImage,
which invalidates skia's cache in its dtor. This means that reusing the
same uniqueID will still result in purging of the decode from skia's
cache.

Also note that the only case where the reuse of uniqueID is important
is for each frame of a multiframe image. For static images,
BitmapImage already ensures caching of the decode by storing the
PaintImage which internally keeps a reference to the SkImage, and is
cleared in DestroyDecodedData. All cases which require animation of
images go through cc and use cc's image decode cache. The few cases,
for instance Canvas2D, which decode on the main thread and use skia's
cache only use the first frame of an animated image.

R=chrishtr@chromium.org, fmalita@chromium.org

Bug: 753639
Cq-Include-Trybots: master.tryserver.blink:linux_trusty_blink_rel
Change-Id: I78dcf84a8dad25c7742526af31c05ee381da7cf4
Reviewed-on: https://chromium-review.googlesource.com/677916
Reviewed-by: Florin Malita <fmalita@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504185}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 653954392a8a6b43036f22fb402ce76a9d55c262