[css] Fix scroll anchor X dirty table layout crash
When an element is selected as scroll anchor, we need to get its
bounding box. But when a table has unlaid out table parts, we DCHECK
that its borders aren't being queried. When a table with unlaid out
table parts is the scroll anchor, the DCHECK was triggered.

The "fix" in this CL is to use the old border widths when scroll
anchoring code is looking for an anchor.

Bug: 746570
Change-Id: I424e3807c2ed7a7a20f02617e2e0586dba83e708
Reviewed-on: https://chromium-review.googlesource.com/679994
Reviewed-by: Steve Kobes <skobes@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504279}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 89aa7bd6d74c9aedbdb8a61048911eafed1474e1