cc/blink: Plumb resetting of image animations from blink to cc.
The animation can be reset to start from the beginning in blink. Plumb
this to cc to replicate this behaviour if image animations are being
ticked by the compositor. This uses a counter sequence incremented on
the main thread each time the animation is reset to act as a
synchronization point, which is necessary since recordings can be
cached and resued by blink.

Bug: 735743
Cq-Include-Trybots: master.tryserver.blink:linux_trusty_blink_rel
Change-Id: I7ff6fe22b1c5179a1ea8ec893f5d239895a4d2a4
Reviewed-on: https://chromium-review.googlesource.com/664245
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504290}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e58f9b7f44eadc1b5456a88939ab4fe861f42b4