Remove StringImpl allocation in AtomicHTMLToken::InitializeAttributes
Uses the AtomicString(Vector<UChar>) constructor for Attribute values.
This is to remove an unnecessary memory allocation and conversion of
attribute values to String (using Value8BitIfNecessary). The allocation
is unnecessary for cases where the value is already present in
AtomicStringTable.

This CL also changes lookupHTMLTag to directly return an AtomicString,
instead of converting it to a StringImpl and then back to an
AtomicString.

Bug: 
Change-Id: I756f99cd57ebc20828d4a0430ba34f1018c730a4
Reviewed-on: https://chromium-review.googlesource.com/685512
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505029}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de41eb4378b73b5ad8f25526f6a3b2222715de2d