[PaintWorklet] Fix geometry when browser zooms
It is discovered in the bug that when the device pixel ratio is 1.7 and
the browser zoom is 1/3, the geometry of the background is 199*199 where
it should be 200*200. There are two reasons:
1. The ApplySubPixelHeuristicToImageSize in BackgroundImageGeometry
   makes the fractional part lost.
2. The conversion from FloatSize to IntSize also has impact.

This CL caches a logical_tile_size_ in the BackgroundImageGeometry, which
is the tile size without applying subpixel heuristics. And then we pass this
size all the way to the CSSPaintDefinition::Paint.

Bug: 683789
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2
Change-Id: I4a24d7afe5408d3fbfbef4b63e3c221f13ed1612
Reviewed-on: https://chromium-review.googlesource.com/678019
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505150}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6d27042dd93aabdddcdddddffefe6a0f7d3a7a1