Added VRSession creation.
This CL only enables the creation of exclusive VRSession objects. It
does not provide the ability to present to a headset, or many of the
functions of a VRSession. Those will be added in follow up CLs to allow
the more complicated logic to be reviewed independently.

Bug: 670510
Change-Id: Ic2b3b1801bf8f3ee4263d009daab8215b0b179d5
Reviewed-on: https://chromium-review.googlesource.com/685396
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505305}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48315306487041b1944ea08c89288933168e9259