Preserve browser controls size for remote-to-local main frame navigations.
The browser controls size might be updated in a resize message that
comes in while the main frame is remote.  Previously,
WebViewImpl::ResizeWithBrowserControls early-returned if the main
frame is remote, saving the main widget size, but ignoring the browser
controls size update.  This CL fixes ResizeWithBrowserControls to also
save the new browser controls size, so that it can be used if the
remote main frame swaps back to a local frame.

This affected URL bar hiding on Android, where the resize message was
sent while there was a provisional main frame that hasn't swapped into
the tree yet, causing the renderer to think that the browser controls
had zero height.

Bug: 769321
Change-Id: I81558bf08483020dc3af7d8a57aa49705704dc05
Reviewed-on: https://chromium-review.googlesource.com/692683
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505693}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a26adbb2e7f97ecb88d7ae9dcf93fe3f8f7d8601