Remove redundant `this` reference in core/dom
There used to be necessity to use explicit `this->` before the Blink
renaming happend, e.g.
  LayoutObject layoutObject = this->layoutObject();
but as we use Chromium style names for local variables and methods,
the above can be
  LayoutObject layout_object = GetLayoutObject();
and `this->` is not necessary.

Change-Id: I789f3457c2825c63e1c16799a887d1778b305646
Reviewed-on: https://chromium-review.googlesource.com/694803
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Hayato Ito <hayato@chromium.org>
Commit-Queue: Takayoshi Kochi <kochi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505953}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5824890b01865469b700c4494434f4f5b8007d38