blink: Assert valid frame indexes used for image generator content id.
When querying for the content id for a frame, the frame index must be
less than the size of frames available in the generator. DCHECK this
and use vector.at, instead of vector[] which has an undefined behaviour
when the access is out of bounds.

R=pdr@chromium.org

Change-Id: Id6300be8898361038bc30838e6218016617d6deb
Reviewed-on: https://chromium-review.googlesource.com/696669
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506093}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e7690536fc52e37216447faa9eb3f17d7e76436