Account for position:sticky offset in composited-ancestor offset calculation.
In the case where compositor is handling sticky offset, we need to remove the
Blink-side offset to avoid double-counting. Previously this was done directly
in CompositedLayerMapping::UpdateMainGraphicsLayerGeometry but that misses cases
such as the position of child_transform_layer_.

In general CompositedLayerMapping should not know about the sticky offset when
the compositor is going to handle it, so this CL removes it from the composited
ancestor offset calculation which then bubbles down to the required locations.

Bug: 762962
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2
Change-Id: I31ce57e6623f0ce3995abbb7d79dccf94022354b
Reviewed-on: https://chromium-review.googlesource.com/665581
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506247}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07d92b832e636397f113353dc9413c247996e250