[CSSParser] Fix handling of unterminated @rules in a declaration block.
Currently, when we encounter something like:

   .a { @extends }

We would consume up the '}', expecting a semicolon. We then erroneously
consume '}' thinking that it's a ';', which would cause the next rule
to be parsed incorrectly.

This patch fixes this issue by only consuming if it is a ';'. We also
add DCHECK to ensure that we never consume a '}' to catch these errors
in the future.

I also manually tested this on the pixel 2 page which has this issue
(see bug) and it seems to work fine.

Bug: 771679
Change-Id: Idbe014f76e32af60af0c18fdcac9d68e95bcfa2c
Reviewed-on: https://chromium-review.googlesource.com/701894
Reviewed-by: nainar <nainar@chromium.org>
Reviewed-by: meade_UTC10 <meade@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506633}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8fe810ac5d111f0a0058e62886eec20ce6694a7