S13nServiceWorker: Don't call OnFallback when FetchEvent was rejected
When FetchEvent.respondWith() is called with a rejecting promise in a service
worker and if OnFetchEventFinished() of ServiceWorkerSubresourceLoader is called
before OnResponse(), OnFallback() passes |url_loader_client_| to the network
loader factory, and the renderer process crashes while calling
OnReceiveResponse().

  OnFetchEventFinished()
   => OnFallback()
     => std::move(url_loader_client_)

  OnResponse()
   => StartResponse()
    => DeliverErrorResponse()
     => CommitResponseHeaders()
      => url_loader_client_->OnReceiveResponse() (* crash here)


OnFetchEventFinished() should not call OnFallback() when |status| is
ERROR_EVENT_WAITUNTIL_REJECTED. And also, we don't need to fallback to the
network because it's for main resource load when FetchEvent failed.

Bug: 715640
Change-Id: I78d31c118ad9169740e3db4c561aa25e2dbae83e
Reviewed-on: https://chromium-review.googlesource.com/700276
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74a832aa58e038b2198c6e881432e1f8be990c8c