Fix flaky test HugeScrollingThumbPosition
Looks like the cause is that the scrollbar margin, which we mock in a
custom WebThemeEngine gets read once when the ScrollbarThemeOverlay is
created and is then cached. However, ScrollbarThemes are static so once
it's created once we never recreate it. Thus, if another test runs
before this one that doesn't use our mocked WebThemeEngine, this test
will use the real margins.

Longer term this will be fixed by making ScrollbarThemes not be global
singletons (being worked on in the context of crbug.com/740879), then we
can just recreate a ScrollbarTheme in this test. For now, we'll just
account for the used margin in our expectation to prevent flakiness.

Bug: 769350
Change-Id: I8e93ac3def151963d8e07cf0bb9cc203f8d51279
Reviewed-on: https://chromium-review.googlesource.com/703491
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507078}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de787538f5df9e7edfd32d92a8795eccbd87a144